+++ src/interpreter.cpp
@@ -47,8 +47,9 @@ using namespace Interpreter;
 
 struct PackedERPreliminary {
     std::shared_ptr<earl::value::Obj> lhs_getter_accessor;
-    PackedERPreliminary(std::shared_ptr<earl::value::Obj> lhs_get = nullptr)
-        : lhs_getter_accessor(lhs_get) {}
+    bool this_;
+    PackedERPreliminary(std::shared_ptr<earl::value::Obj> lhs_get = nullptr, bool this_ = false)
+        : lhs_getter_accessor(lhs_get), this_(this_) {}
 };
 
 static std::shared_ptr<earl::value::Obj>
@@ -89,12 +90,12 @@ eval_stmt_let_wcustom_buffer(StmtLet *stmt,
         value = unpack_ER(rhs, ctx, _ref);
 
     if (stmt->m_id->lexeme() == "_")
-        return nullptr;
+        return std::make_shared<earl::value::Void>();
 
     std::shared_ptr<earl::variable::Obj> var
         = std::make_shared<earl::variable::Obj>(stmt->m_id.get(), value, stmt->m_attrs);
     ctx->variable_add(var);
-    return nullptr;
+    return std::make_shared<earl::value::Void>();
 }
 
 static std::shared_ptr<earl::value::Obj>
@@ -313,6 +314,10 @@ unpack_ER(ER &er, std::shared_ptr<Ctx> &ctx, bool ref, PackedERPreliminary *perp
             return Intrinsics::call(er.id, params, ctx);
         if (er.is_member_intrinsic()) {
             assert(perp && perp->lhs_getter_accessor);
+            if (!Intrinsics::is_member_intrinsic(er.id, static_cast<int>(perp->lhs_getter_accessor->type()))) {
+                ERR_WARGS(Err::Type::Fatal, "type `%s` does not implement member intrinsic `%s`",
+                          earl::value::type_to_str(perp->lhs_getter_accessor->type()).c_str(), er.id.c_str());
+            }
             return Intrinsics::call_member(er.id,
                                            perp->lhs_getter_accessor->type(),
                                            perp->lhs_getter_accessor,
@@ -320,8 +325,14 @@ unpack_ER(ER &er, std::shared_ptr<Ctx> &ctx, bool ref, PackedERPreliminary *perp
                                            ctx);
         }
 
-        if (ctx->type() == CtxType::Class)
+        if (ctx->type() == CtxType::Class) {
+            std::shared_ptr<earl::function::Obj> func = nullptr;
+            if (ctx->function_exists(er.id))
+                func = ctx->function_get(er.id);
+            if ((!perp || !perp->this_) && (func && er.ctx != ctx && !func->is_pub()))
+                ERR_WARGS(Err::Type::Fatal, "member variable `%s` is missing the @pub attribute", func->id().c_str());
             return eval_user_defined_function(er.id, params, ctx);
+        }
 
         // We need to have this function to gen the parameters so we
         // know which ones need to be taken as a reference. NOTE: The
@@ -333,6 +344,8 @@ unpack_ER(ER &er, std::shared_ptr<Ctx> &ctx, bool ref, PackedERPreliminary *perp
     else if (er.is_ident()) {
         if (ctx->variable_exists(er.id)) {
             auto var = ctx->variable_get(er.id);
+            if ((!perp || !perp->this_) && (er.ctx != ctx && !var->is_pub()))
+                ERR_WARGS(Err::Type::Fatal, "member variable `%s` is missing the @pub attribute", var->id().c_str());
             if (!ref)
                 return var->value()->copy();
             return var->value();
@@ -523,7 +536,8 @@ eval_expr_term_get(ExprGet *expr, std::shared_ptr<Ctx> &ctx, bool ref) {
         if (!fctx->in_class())
             ERR(Err::Type::Fatal, "Must be in a class context when using the `this` keyword");
 
-        auto value = unpack_ER(right_er, fctx->get_outer_class_owner_ctx(), /*ref=*/true);
+        PackedERPreliminary perp(nullptr, true);
+        auto value = unpack_ER(right_er, fctx->get_outer_class_owner_ctx(), /*ref=*/true, /*perp=*/&perp);
         return ER(value, ERT::Literal);
     }
     else {
@@ -579,8 +593,12 @@ eval_expr_term_array_access(ExprArrayAccess *expr, std::shared_ptr<Ctx> &ctx, bo
         auto str = dynamic_cast<earl::value::Str *>(left_value.get());
         return ER(str->nth(idx_value), ERT::Literal);
     }
+    else if (left_value->type() == earl::value::Type::Tuple) {
+        auto tuple = dynamic_cast<earl::value::Tuple *>(left_value.get());
+        return ER(tuple->nth(idx_value), static_cast<ERT>(ERT::Literal|ERT::TupleAccess));
+    }
     else
-        ERR(Err::Type::Fatal, "cannot use `[]` on non-list or non-str type");
+        ERR(Err::Type::Fatal, "cannot use `[]` on non-list, non-tuple, or non-str type");
 }
 
 static ER
@@ -645,6 +663,17 @@ eval_expr_term_range(ExprRange *expr, std::shared_ptr<Ctx> &ctx, bool ref) {
     }
 }
 
+static ER
+eval_expr_term_tuple(ExprTuple *expr, std::shared_ptr<Ctx> &ctx, bool ref) {
+    std::vector<std::shared_ptr<earl::value::Obj>> values = {};
+    for (auto &e : expr->m_exprs) {
+        ER er = Interpreter::eval_expr(e.get(), ctx, ref);
+        auto value = unpack_ER(er, ctx, ref);
+        values.push_back(value);
+    }
+    return ER(std::make_shared<earl::value::Tuple>(values), ERT::Literal);
+}
+
 ER
 eval_expr_term(ExprTerm *expr, std::shared_ptr<Ctx> &ctx, bool ref) {
     switch (expr->get_term_type()) {
@@ -662,7 +691,7 @@ eval_expr_term(ExprTerm *expr, std::shared_ptr<Ctx> &ctx, bool ref) {
     case ExprTermType::None:          return eval_expr_term_none(dynamic_cast<ExprNone *>(expr));
     case ExprTermType::Closure:       return eval_expr_term_closure(dynamic_cast<ExprClosure *>(expr), ctx, ref);
     case ExprTermType::Range:         return eval_expr_term_range(dynamic_cast<ExprRange *>(expr), ctx, ref);
-    case ExprTermType::Tuple:         UNIMPLEMENTED("ExprTermType::Tuple");
+    case ExprTermType::Tuple:         return eval_expr_term_tuple(dynamic_cast<ExprTuple *>(expr), ctx, ref);
     default:                          ERR_WARGS(Err::Type::Fatal, "unknown term: `%d`", (int)expr->get_term_type());
     }
     assert(false && "unreachable");
@@ -740,12 +769,12 @@ eval_stmt_let(StmtLet *stmt, std::shared_ptr<Ctx> &ctx) {
         value = unpack_ER(rhs, ctx, ref);
 
     if (stmt->m_id->lexeme() == "_")
-        return nullptr;
+        return std::make_shared<earl::value::Void>();
 
     std::shared_ptr<earl::variable::Obj> var
         = std::make_shared<earl::variable::Obj>(stmt->m_id.get(), value, stmt->m_attrs);
     ctx->variable_add(var);
-    return nullptr;
+    return std::make_shared<earl::value::Void>();
 }
 
 std::shared_ptr<earl::value::Obj>
@@ -819,6 +848,10 @@ std::shared_ptr<earl::value::Obj>
 eval_stmt_mut(StmtMut *stmt, std::shared_ptr<Ctx> &ctx) {
     ER left_er = Interpreter::eval_expr(stmt->m_left.get(), ctx, true);
     ER right_er = Interpreter::eval_expr(stmt->m_right.get(), ctx, false);
+
+    if (left_er.is_tuple_access())
+        ERR(Err::Type::Fatal, "cannot mutate tuple type as they are immutable");
+
     auto l = unpack_ER(left_er, ctx, true);
     auto r = unpack_ER(right_er, ctx, false);
     switch (stmt->m_equals->type()) {
@@ -899,6 +932,29 @@ eval_stmt_foreach(StmtForeach *stmt, std::shared_ptr<Ctx> &ctx) {
         }
         ctx->variable_remove(enumerator->id());
     }
+    else if (expr->type() == earl::value::Type::Tuple) {
+        auto tuple = std::dynamic_pointer_cast<earl::value::Tuple>(expr);
+        if (tuple->value().size() == 0)
+            return result;
+        auto enumerator = std::make_shared<earl::variable::Obj>(stmt->m_enumerator.get(), tuple->value()[0]);
+        if (ctx->variable_exists(enumerator->id())) {
+            Err::err_wtok(stmt->m_enumerator.get());
+            ERR_WARGS(Err::Type::Redeclared, "variable `%s` is already declared", stmt->m_enumerator->lexeme().c_str());
+        }
+        ctx->variable_add(enumerator);
+        for (size_t i = 0; i < tuple->value().size(); ++i) {
+            if (i != 0)
+                enumerator->reset(tuple->value()[i]);
+            result = Interpreter::eval_stmt_block(stmt->m_block.get(), ctx);
+            if (result && result->type() == earl::value::Type::Break) {
+                result = nullptr;
+                break;
+            }
+            if (result && result->type() != earl::value::Type::Void)
+                break;
+        }
+        ctx->variable_remove(enumerator->id());
+    }
     else if (expr->type() == earl::value::Type::Str) {
         auto str = std::dynamic_pointer_cast<earl::value::Str>(expr);
         if (str->value().size() == 0)
@@ -922,7 +978,7 @@ eval_stmt_foreach(StmtForeach *stmt, std::shared_ptr<Ctx> &ctx) {
         }
     }
     else
-        ERR(Err::Type::Fatal, "unable to perform a `for` loop with an expression other than a list type");
+        ERR(Err::Type::Fatal, "unable to perform a `for` loop with an expression other than a list, str, or tuple type");
 
     return result;
 }
@@ -1137,7 +1193,7 @@ eval_stmt_enum(StmtEnum *stmt, std::shared_ptr<Ctx> &ctx) {
     auto _enum = std::make_shared<earl::value::Enum>(stmt, std::move(elems), stmt->m_attrs);
     assert(ctx->type() == CtxType::World);
     dynamic_cast<WorldCtx *>(ctx.get())->enum_add(std::move(_enum));
-    return nullptr;
+    return std::make_shared<earl::value::Void>();
 }
 
 std::shared_ptr<earl::value::Obj>